requirements.txt
asgiref==3.8.1
Django~=5.2b1
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
djangorestframework~=3.15.2
black==23.12.1
pylint==2.17.5
Pillow>=8.0.0
drf-yasg==1.21.7
pandas~=2.2.3
django-cors-headers
selenium==4.15.2
webdriver-manager==4.0.1
pytest~=8.3.5
pytest-django~=4.5.2